requirements.txt
astroid==2.1.0
dj-database-url==0.5.0
django-filter=2.0.0
djangorestframework=3.9.0
Django==2.1.4
gunicorn==19.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
markdown=3.0.1
mccabe==0.6.1
psycopg2-binary==2.7.6.1
pylint==2.2.2
pytz==2018.7
six==1.12.0
whitenoise==3.3.0
wrapt==1.10.11
